Wrap consensus_encoding tests and examples
The tests and examples require rustc 1.65 or later.

Move the tests and examples to a `rust_v_1_65` subdirectory and wrap
them in a `#[cfg(rust_v_1_65)]` to allow building with MSRV.
