+++ consensus_encoding/tests/rust_v_1_65/wrappers.rs
@@ -1,0 +1,132 @@
+// Test using wrapper types as we expect the lib to be used.
+
+use bitcoin_consensus_encoding as encoding;
+use encoding::{ArrayEncoder, BytesEncoder, CompactSizeEncoder, Encodable, Encoder2, SliceEncoder};
+
+encoding::encoder_newtype_exact! {
+    /// An encoder that uses an inner `ArrayEncoder`.
+    pub struct TestArrayEncoder<'e>(ArrayEncoder<4>);
+}
+
+encoding::encoder_newtype_exact! {
+    /// An encoder that uses an inner `BytesEncoder`.
+    pub struct TestBytesEncoder<'e>(BytesEncoder<'e>);
+}
+
+#[test]
+fn array_encoder() {
+    #[derive(Debug, Default, Clone)]
+    pub struct Test(u32);
+
+    impl Encodable for Test {
+        type Encoder<'e> = TestArrayEncoder<'e>;
+        fn encoder(&self) -> Self::Encoder<'_> {
+            TestArrayEncoder::new(ArrayEncoder::without_length_prefix(self.0.to_le_bytes()))
+        }
+    }
+
+    let t = Test(0xcafe_babe); // Encodes using an array.
+
+    let want = [0xbe, 0xba, 0xfe, 0xca];
+    let got = encoding::encode_to_vec(&t);
+
+    assert_eq!(got, want);
+}
+
+#[test]
+fn bytes_encoder_without_length_prefix() {
+    #[derive(Debug, Default, Clone)]
+    pub struct Test(Vec<u8>);
+
+    impl Encodable for Test {
+        type Encoder<'e>
+            = TestBytesEncoder<'e>
+        where
+            Self: 'e;
+
+        fn encoder(&self) -> Self::Encoder<'_> {
+            TestBytesEncoder::new(BytesEncoder::without_length_prefix(self.0.as_ref()))
+        }
+    }
+
+    let t = Test(vec![0xca, 0xfe]);
+
+    let want = [0xca, 0xfe];
+    let got = encoding::encode_to_vec(&t);
+
+    assert_eq!(got, want);
+}
+
+#[test]
+fn two_encoder() {
+    #[derive(Debug, Default, Clone)]
+    pub struct Test {
+        a: Vec<u8>,
+        b: Vec<u8>,
+    }
+
+    impl Encodable for Test {
+        type Encoder<'e> = Encoder2<TestBytesEncoder<'e>, TestBytesEncoder<'e>>;
+
+        fn encoder(&self) -> Self::Encoder<'_> {
+            let a = TestBytesEncoder::new(BytesEncoder::without_length_prefix(self.a.as_ref()));
+            let b = TestBytesEncoder::new(BytesEncoder::without_length_prefix(self.b.as_ref()));
+
+            Encoder2::new(a, b)
+        }
+    }
+
+    let t = Test { a: vec![0xca, 0xfe], b: (vec![0xba, 0xbe]) };
+
+    let want = [0xca, 0xfe, 0xba, 0xbe];
+    let got = encoding::encode_to_vec(&t);
+
+    assert_eq!(got, want);
+}
+
+#[test]
+fn slice_encoder() {
+    #[derive(Debug, Default, Clone)]
+    pub struct Test(Vec<Inner>);
+
+    encoding::encoder_newtype! {
+        /// An encoder that uses an inner `SliceEncoder`.
+        pub struct TestEncoder<'e>(Encoder2<CompactSizeEncoder, SliceEncoder<'e, Inner>>);
+    }
+
+    impl Encodable for Test {
+        type Encoder<'e>
+            = TestEncoder<'e>
+        where
+            Self: 'e;
+
+        fn encoder(&self) -> Self::Encoder<'_> {
+            TestEncoder::new(Encoder2::new(
+                CompactSizeEncoder::new(self.0.len()),
+                SliceEncoder::without_length_prefix(&self.0),
+            ))
+        }
+    }
+
+    #[derive(Debug, Default, Clone)]
+    pub struct Inner(u32);
+
+    encoding::encoder_newtype_exact! {
+        /// The encoder for the [`Inner`] type.
+        pub struct InnerArrayEncoder<'e>(ArrayEncoder<4>);
+    }
+
+    impl Encodable for Inner {
+        type Encoder<'e> = InnerArrayEncoder<'e>;
+        fn encoder(&self) -> Self::Encoder<'_> {
+            // Big-endian to make reading the test assertion easier.
+            InnerArrayEncoder::new(ArrayEncoder::without_length_prefix(self.0.to_be_bytes()))
+        }
+    }
+
+    let t = Test(vec![Inner(0xcafe_babe), Inner(0xdead_beef)]);
+    let encoded = encoding::encode_to_vec(&t);
+
+    let want = [0x02, 0xca, 0xfe, 0xba, 0xbe, 0xde, 0xad, 0xbe, 0xef];
+    assert_eq!(encoded, want);
+}
